Rearrange D&D table rows
- added new DnD-class
- rename RowDnD to TableRowDnD
