--- JAG3D/src/org/applied_geodesy/jag3d/ui/dnd/RowDnD.java
@@ -1,49 +1,0 @@
-/***********************************************************************
-* Copyright by Michael Loesler, https://software.applied-geodesy.org   *
-*                                                                      *
-* This program is free software; you can redistribute it and/or modify *
-* it under the terms of the GNU General Public License as published by *
-* the Free Software Foundation; either version 3 of the License, or    *
-* at your option any later version.                                    *
-*                                                                      *
-* This program is distributed in the hope that it will be useful,      *
-* but WITHOUT ANY WARRANTY; without even the implied warranty of       *
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        *
-* GNU General Public License for more details.                         *
-*                                                                      *
-* You should have received a copy of the GNU General Public License    *
-* along with this program; if not, see <http://www.gnu.org/licenses/>  *
-* or write to the                                                      *
-* Free Software Foundation, Inc.,                                      *
-* 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.            *
-*                                                                      *
-***********************************************************************/
-
-package org.applied_geodesy.jag3d.ui.dnd;
-
-import java.io.Serializable;
-
-public abstract class RowDnD implements Serializable {
-	private static final long serialVersionUID = 7702551810226354087L;
-	private int id = -1, groupId = -1;
-	private boolean enable = true;
-	
-	public int getId() {
-		return id;
-	}
-	void setId(int id) {
-		this.id = id;
-	}
-	public int getGroupId() {
-		return groupId;
-	}
-	void setGroupId(int groupId) {
-		this.groupId = groupId;
-	}
-	public boolean isEnable() {
-		return enable;
-	}
-	void setEnable(boolean enable) {
-		this.enable = enable;
-	}
-}
