+++ JAG3D/src/org/applied_geodesy/jag3d/ui/dnd/TableRowDnD.java
@@ -1,0 +1,41 @@
+/***********************************************************************
+* Copyright by Michael Loesler, https://software.applied-geodesy.org   *
+*                                                                      *
+* This program is free software; you can redistribute it and/or modify *
+* it under the terms of the GNU General Public License as published by *
+* the Free Software Foundation; either version 3 of the License, or    *
+* at your option any later version.                                    *
+*                                                                      *
+* This program is distributed in the hope that it will be useful,      *
+* but WITHOUT ANY WARRANTY; without even the implied warranty of       *
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        *
+* GNU General Public License for more details.                         *
+*                                                                      *
+* You should have received a copy of the GNU General Public License    *
+* along with this program; if not, see <http://www.gnu.org/licenses/>  *
+* or write to the                                                      *
+* Free Software Foundation, Inc.,                                      *
+* 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.            *
+*                                                                      *
+***********************************************************************/
+
+package org.applied_geodesy.jag3d.ui.dnd;
+
+public abstract class TableRowDnD extends DnD {
+	private static final long serialVersionUID = 3983696389688227774L;
+	private int id = -1;
+	private boolean enable = true;
+	
+	public int getId() {
+		return this.id;
+	}
+	void setId(int id) {
+		this.id = id;
+	}
+	public boolean isEnable() {
+		return this.enable;
+	}
+	void setEnable(boolean enable) {
+		this.enable = enable;
+	}
+}
